pyproject.toml
[tool.poetry]
name = "corona-telegram"
version = "0.1.0"
description = ""
authors = ["oryon-dominik <oryon@cyberise.de>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.4.2"
python-dotenv = "^0.12.0"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
